Balkanization
- Both custom classes (PointCompute and Rotor) have been split off into separate files
- `boring_math.py` was replaced with `utils.py`
- `report()` was moved to `utils.py`
- `compute_period()` was removed and the logic was brought into `PointCompute.__init__()`
- `compute_draw_loc()` and `interpolate()` were both moved into `pointcompute.py`
- `PointCompute.compute_segment()` is now a standalone function
- X/Y max/min are now updated from the results of `compute_segment()` rather than on each point calculation
- Added `utils.fl_eq()` to clean up floating point comparison checks
- `PointCompute.tuples()` has been replaced with making PointCompute iterable
